\name{greetings}
\alias{greetings}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
greetings
}
\description{
saluda a alguien
}
\usage{
version(1.0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
}
\value{
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
greetings("Pepe")
[1] "Saludo Pepe"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
